+++ source/renderer/app/actions/ada/wallets-actions.js
@@ -13,13 +13,13 @@ export type WalletImportFromFileParams = {
 export default class WalletsActions {
   createWallet: Action<{ name: string, password: ?string }> = new Action();
   // eslint-disable-next-line max-len
-  restoreWallet: Action<{recoveryPhrase: string, walletName: string, walletPassword: ?string }> = new Action();
+  restoreWallet: Action<{recoveryPhrase: string, walletName: string, walletPassword: ?string, type: string, certificatePassword?: string }> = new Action();
   importWalletFromFile: Action<WalletImportFromFileParams> = new Action();
   deleteWallet: Action<{ walletId: string }> = new Action();
   sendMoney: Action<{ receiver: string, amount: string, password: ?string }> = new Action();
   chooseWalletExportType: Action<{ walletExportType: walletExportTypeChoices }> = new Action();
   // eslint-disable-next-line max-len
-  generateCertificate: Action<{ password: string, repeatPassword: string, intl: Object, filePath: string }> = new Action();
+  generateCertificate: Action<{ password: string, repeatPassword: string, filePath: string }> = new Action();
   verifyCertificate: Action<{ recoveryPhrase: Array<string>, password: string }> = new Action();
   updateCertificateStep: Action<any> = new Action();
   closeCertificateGeneration: Action<any> = new Action();