+++ source/renderer/app/api/common.js
@@ -68,6 +68,7 @@ export type GetLocalTimeDifferenceResponse = number;
 export type GetWalletsResponse = Array<Wallet>;
 export type GetWalletRecoveryPhraseResponse = Array<string>;
 export type GetWalletCertificateRecoveryPhraseResponse = Array<string>;
+export type GetWalletRecoveryPhraseFromCertificateResponse = Array<string>;
 export type RestoreWalletResponse = Wallet;
 export type UpdateWalletResponse = Wallet;
 export type UpdateWalletPasswordResponse = boolean;