+++ source/renderer/app/components/wallet/paper-wallet-certificate/VerificationDialog.js
@@ -1,6 +1,6 @@
 // @flow
 import React, { Component } from 'react';
-import { join, isEqual } from 'lodash';
+import { join } from 'lodash';
 import SvgInline from 'react-svg-inline';
 import { observer } from 'mobx-react';
 import classnames from 'classnames';
@@ -17,7 +17,6 @@ import { InvalidMnemonicError } from '../../../i18n/errors';
 import globalMessages from '../../../i18n/global-messages';
 import showPasswordIcon from '../../../assets/images/show-pass-ic.inline.svg';
 import hidePasswordIcon from '../../../assets/images/hide-pass-ic.inline.svg';
-
 import styles from './VerificationDialog.scss';
 
 const messages = defineMessages({
@@ -97,7 +96,6 @@ type Props = {
 };
 
 @observer
-// eslint-disable-next-line
 export default class VerificationDialog extends Component<Props, State> {
 
   static contextTypes = {
@@ -110,7 +108,9 @@ export default class VerificationDialog extends Component<Props, State> {
     showPassword: false,
     isPasswordValid: false,
     isRecoveryPhraseValid: false,
-  }
+  };
+
+  recoveryPhraseAutocomplete: Autocomplete;
 
   form = new ReactToolboxMobxForm({
     fields: {
@@ -127,10 +127,10 @@ export default class VerificationDialog extends Component<Props, State> {
 
           const value = join(field.value, ' ');
           if (value === '') return [false, this.context.intl.formatMessage(globalMessages.fieldIsRequired)];
-          const isRecoveryPhraseValid = isEqual(walletCertificateRecoveryPhrase, field.value);
+          const isRecoveryPhraseValid = walletCertificateRecoveryPhrase === value;
           this.setState({
             isRecoveryPhraseValid,
-            // uncheck confirmation boxes if recovery phrase is not valid and mark as disabled
+            // disabled and uncheck confirmation checkboxes if recovery phrase is not valid
             storingConfirmed: isRecoveryPhraseValid ? storingConfirmed : false,
             recoveringConfirmed: isRecoveryPhraseValid ? recoveringConfirmed : false,
           });
@@ -138,7 +138,6 @@ export default class VerificationDialog extends Component<Props, State> {
             isRecoveryPhraseValid,
             this.context.intl.formatMessage(new InvalidMnemonicError())
           ];
-
         }],
       },
       password: {
@@ -155,7 +154,7 @@ export default class VerificationDialog extends Component<Props, State> {
           const isPasswordValid = this.props.walletCertificatePassword === field.value;
           this.setState({
             isPasswordValid,
-            // uncheck confirmation boxes if password is not valid and mark as disabled
+            // disabled and uncheck confirmation checkboxes if recovery phrase is not valid
             storingConfirmed: isPasswordValid ? storingConfirmed : false,
             recoveringConfirmed: isPasswordValid ? recoveringConfirmed : false,
           });
@@ -190,13 +189,19 @@ export default class VerificationDialog extends Component<Props, State> {
 
   resetForm = () => {
     const { form } = this;
-    form.$('password').reset();
-    form.$('recoveryPhrase').reset();
+    // Cancel all debounced field validations
+    form.each((field) => { field.debouncedValidation.cancel(); });
+    form.reset();
+    form.showErrors(false);
+
+    // Autocomplete has to be reset manually
+    this.recoveryPhraseAutocomplete.clear();
+
     this.setState({
       storingConfirmed: false,
       recoveringConfirmed: false,
     });
-  }
+  };
 
   render() {
     const { intl } = this.context;
@@ -218,12 +223,24 @@ export default class VerificationDialog extends Component<Props, State> {
       'verificationDialog',
     ]);
 
+    const storingUnderstandanceCheckboxClasses = classnames([
+      styles.checkbox,
+      'storingUnderstandance',
+    ]);
+
+    const recoveringUnderstandanceCheckboxClasses = classnames([
+      styles.checkbox,
+      'recoveringUnderstandance'
+    ]);
+
     const actions = [
       {
+        className: 'clearButton',
         label: intl.formatMessage(messages.clearButtonLabel),
         onClick: resetForm.bind(this),
       },
       {
+        className: 'continueButton',
         label: intl.formatMessage(globalMessages.dialogButtonContinueLabel),
         primary: true,
         disabled: !storingConfirmed || !recoveringConfirmed,
@@ -246,6 +263,7 @@ export default class VerificationDialog extends Component<Props, State> {
               className={styles.recoveryPhrase}
               options={suggestedMnemonics}
               maxSelections={15}
+              ref={(autocomplete) => { this.recoveryPhraseAutocomplete = autocomplete; }}
               {...recoveryPhraseField.bind()}
               error={recoveryPhraseField.error}
               maxVisibleOptions={5}
@@ -269,7 +287,7 @@ export default class VerificationDialog extends Component<Props, State> {
             </div>
 
             <Checkbox
-              className={styles.checkbox}
+              className={storingUnderstandanceCheckboxClasses}
               label={intl.formatMessage(messages.storingUnderstandanceLabel)}
               onChange={this.onStoringConfirmationChange.bind(this)}
               checked={storingConfirmed}
@@ -278,7 +296,7 @@ export default class VerificationDialog extends Component<Props, State> {
             />
 
             <Checkbox
-              className={styles.checkbox}
+              className={recoveringUnderstandanceCheckboxClasses}
               label={intl.formatMessage(messages.recoveringUnderstandanceLabel)}
               onChange={this.onRecoveringConfirmationChange.bind(this)}
               checked={recoveringConfirmed}
@@ -298,7 +316,7 @@ export default class VerificationDialog extends Component<Props, State> {
 
   onStoringConfirmationChange = () => {
     this.setState({
-      storingConfirmed: !this.state.storingConfirmed
+      storingConfirmed: !this.state.storingConfirmed,
     });
   };
 
@@ -308,7 +326,7 @@ export default class VerificationDialog extends Component<Props, State> {
 
   onRecoveringConfirmationChange = () => {
     this.setState({
-      recoveringConfirmed: !this.state.recoveringConfirmed
+      recoveringConfirmed: !this.state.recoveringConfirmed,
     });
   };
 }