+++ source/renderer/app/stores/ada/AdaWalletsStore.js
@@ -1,10 +1,11 @@
 // @flow
 import { observable, action, runInAction } from 'mobx';
-import { get, difference } from 'lodash';
+import { get } from 'lodash';
 import { encryptPassphrase } from '../../api/ada/lib/encryptPassphrase';
 import WalletStore from '../WalletStore';
 import Wallet from '../../domain/Wallet';
 import { matchRoute, buildRoute } from '../../utils/routing';
+import { i18nContext } from '../../utils/i18nContext';
 import Request from '.././lib/LocalizedRequest';
 import { ROUTES } from '../../routes-config';
 import WalletAddDialog from '../../components/wallet/WalletAddDialog';
@@ -16,6 +17,7 @@ import type {
   CreateTransactionResponse, CreateWalletResponse, DeleteWalletResponse,
   GetWalletsResponse, RestoreWalletResponse,
   GetWalletRecoveryPhraseResponse, GetWalletCertificateRecoveryPhraseResponse,
+  GetWalletRecoveryPhraseFromCertificateResponse,
 } from '../../api/common';
 
 export default class AdaWalletsStore extends WalletStore {
@@ -30,6 +32,7 @@ export default class AdaWalletsStore extends WalletStore {
   @observable sendMoneyRequest: Request<CreateTransactionResponse> = new Request(this.api.ada.createTransaction);
   @observable getWalletRecoveryPhraseRequest: Request<GetWalletRecoveryPhraseResponse> = new Request(this.api.ada.getWalletRecoveryPhrase);
   @observable getWalletCertificateRecoveryPhraseRequest: Request<GetWalletCertificateRecoveryPhraseResponse> = new Request(this.api.ada.getWalletCertificateRecoveryPhrase);
+  @observable getWalletRecoveryPhraseFromCertificateRequest: Request<GetWalletRecoveryPhraseFromCertificateResponse> = new Request(this.api.ada.getWalletRecoveryPhraseFromCertificate);
   @observable restoreRequest: Request<RestoreWalletResponse> = new Request(this.api.ada.restoreWallet);
   /* eslint-enable max-len */
 
@@ -87,6 +90,10 @@ export default class AdaWalletsStore extends WalletStore {
 
   isValidMnemonic = (mnemonic: string) => this.api.ada.isValidMnemonic(mnemonic);
 
+  isValidCertificateMnemonic = (
+    mnemonic: string,
+  ) => this.api.ada.isValidCertificateMnemonic(mnemonic);
+
   // TODO - call endpoint to check if private key is valid
   isValidPrivateKey = () => { return true; }; // eslint-disable-line
 
@@ -127,7 +134,30 @@ export default class AdaWalletsStore extends WalletStore {
     recoveryPhrase: string,
     walletName: string,
     walletPassword: ?string,
+    type: string,
+    certificatePassword?: string,
   }) => {
+    const data = {
+      recoveryPhrase: params.recoveryPhrase,
+      walletName: params.walletName,
+      walletPassword: params.walletPassword,
+    };
+
+    if (params.type === 'certificate') {
+      // Unscramble 15-word wallet certificate mnemonic to 12-word mnemonic
+      const unscrambledRecoveryPhrase: ?GetWalletCertificateRecoveryPhraseResponse = await (
+        this.getWalletRecoveryPhraseFromCertificateRequest.execute({
+          passphrase: params.certificatePassword,
+          scrambledInput: params.recoveryPhrase,
+        }).promise
+      );
+      if (unscrambledRecoveryPhrase) {
+        data.recoveryPhrase = unscrambledRecoveryPhrase;
+      } else {
+        throw new Error('Invalid mnemonic');
+      }
+    }
+
     this.restoreRequest.reset();
     this._setIsRestoreActive(true);
     // Hide restore wallet dialog some time after restore has been started
@@ -137,7 +167,7 @@ export default class AdaWalletsStore extends WalletStore {
       if (!this.restoreRequest.isError) this._toggleAddWalletDialogOnActiveRestoreOrImport();
     }, this.WAIT_FOR_SERVER_ERROR_TIME);
 
-    const restoredWallet = await this.restoreRequest.execute(params).promise;
+    const restoredWallet = await this.restoreRequest.execute(data).promise;
     setTimeout(() => {
       this._setIsRestoreActive(false);
       this.actions.dialogs.closeActiveDialog.trigger();
@@ -208,7 +238,6 @@ export default class AdaWalletsStore extends WalletStore {
   _generateCertificate = async (params: {
     password: string,
     repeatPassword: string,
-    intl: Object,
     filePath: string,
   }) => {
     try {
@@ -233,7 +262,9 @@ export default class AdaWalletsStore extends WalletStore {
           input: recoveryPhrase.join(' '),
         }).promise
       );
-      this.walletCertificateRecoveryPhrase = walletCertificateRecoveryPhrase;
+      this.walletCertificateRecoveryPhrase = walletCertificateRecoveryPhrase
+        ? walletCertificateRecoveryPhrase.join(' ')
+        : '';
 
       // Generate random spending password
       const spendingPassword = encryptPassphrase(`${params.password}-${Date.now()}`);
@@ -268,7 +299,6 @@ export default class AdaWalletsStore extends WalletStore {
       this._downloadCertificate(
         walletAddress,
         walletCertificateRecoveryPhrase,
-        params.intl,
         params.filePath,
       );
     } catch (error) {
@@ -281,10 +311,12 @@ export default class AdaWalletsStore extends WalletStore {
   _downloadCertificate = action((
     address: string,
     recoveryPhrase: Array<string>,
-    intl: Object,
     filePath: string,
   ) => {
     setTimeout(() => { // Timeout is used to allow enought time for button text re-rendering
+      const locale = this.stores.profile.currentLocale;
+      const intl = i18nContext(locale);
+
       downloadPaperWalletCertificate({
         address,
         mnemonics: recoveryPhrase,
@@ -313,13 +345,8 @@ export default class AdaWalletsStore extends WalletStore {
     password: string,
   }) => {
     const { recoveryPhrase, password } = params;
-
-    const recoveryPhraseMatch = difference(
-      this.walletCertificateRecoveryPhrase, recoveryPhrase
-    ).length === 0;
-
     const passwordMatch = this.walletCertificatePassword === password;
-    if (!recoveryPhraseMatch || !passwordMatch) {
+    if ((this.walletCertificateRecoveryPhrase !== recoveryPhrase.join(' ')) || !passwordMatch) {
       this.walletCertificateHasError = true;
     } else {
       this.walletCertificateHasError = false;